fix all target build settings
