+++ HoaryFox/RH7/HoaryFoxRH7.csproj
@@ -9,6 +9,7 @@
     <GenerateResourceUsePreserializedResources>true</GenerateResourceUsePreserializedResources>
     <AssemblyName>HoaryFox</AssemblyName>
     <RunAnalyzersDuringBuild>false</RunAnalyzersDuringBuild>
+    <RootNamespace>HoaryFox</RootNamespace>
   </PropertyGroup>
   
   <ItemGroup>
@@ -18,21 +19,6 @@
   
   <ItemGroup>
     <ProjectReference Include="..\Common\HoaryFoxCommon.csproj" />
-  </ItemGroup>
-  
-  <ItemGroup>
-    <Compile Update="Properties\Resource.Designer.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>Resource.resx</DependentUpon>
-    </Compile>
-  </ItemGroup>
-  
-  <ItemGroup>
-    <EmbeddedResource Update="Properties\Resource.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>Resource.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
   </ItemGroup>
   
   <PropertyGroup Condition="$(Configuration) == 'Debug' AND $([MSBuild]::IsOSPlatform(Windows))">
