--- KarambaConnect/RH6_K2/karambaConnectRH6K2.csproj
@@ -1,45 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-	
-  <PropertyGroup>
-    <TargetFramework>net48</TargetFramework>
-    <Version>2.0.2</Version>
-    <Title>karambaConnect</Title>
-    <Description>HoaryFox karamba connect extension</Description>
-    <TargetExt>.gha</TargetExt>
-    <Authors>hrntsm</Authors>
-    <EnforceCodeStyleInBuild>false</EnforceCodeStyleInBuild>
-    <EnableNETAnalyzers>false</EnableNETAnalyzers>
-    <GenerateResourceUsePreserializedResources>true</GenerateResourceUsePreserializedResources>
-    <RunAnalyzersDuringBuild>false</RunAnalyzersDuringBuild>
-    <PackageId>karambaConnect</PackageId>
-    <AssemblyName>karambaConnect</AssemblyName>
-    <RootNamespace>karambaConnect</RootNamespace>
-  </PropertyGroup>
-  
-  <ItemGroup>
-    <PackageReference Include="Grasshopper" Version="6.34.21034.7001" IncludeAssets="compile;build" />
-    <PackageReference Include="STBDotNet" Version="0.2.1" />
-  </ItemGroup>
-  
-  <ItemGroup>
-    <ProjectReference Include="..\Common\KarambaConnectCommon.csproj" />
-  </ItemGroup>
-  
-  <ItemGroup>
-    <Reference Include="KarambaCommon">
-      <HintPath>..\Common\Library\RH6_Karamba2\karambaCommon.dll</HintPath>
-      <Private>false</Private>
-    </Reference>
-    <Reference Include="KarambaGH">
-      <HintPath>..\Common\Library\RH6_Karamba2\karamba.gha</HintPath>
-      <Private>false</Private>
-    </Reference>
-  </ItemGroup>
-  
-  <PropertyGroup Condition="$(Configuration) == 'Debug' AND $([MSBuild]::IsOSPlatform(Windows))">
-    <StartProgram>C:\Program Files\Rhino 6\System\Rhino.exe</StartProgram>
-    <StartArguments></StartArguments>
-    <StartAction>Program</StartAction>
-  </PropertyGroup>
-
-</Project>
