+++ KarambaConnect/RH7K1/karambaConnectRH7K1.csproj
@@ -1,0 +1,51 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net48</TargetFramework>
+    <Version>2.0.2</Version>
+    <Title>karambaConnect</Title>
+    <Description>HoaryFox karamba connect extension</Description>
+    <TargetExt>.gha</TargetExt>
+    <Authors>hrntsm</Authors>
+    <EnforceCodeStyleInBuild>false</EnforceCodeStyleInBuild>
+    <EnableNETAnalyzers>false</EnableNETAnalyzers>
+    <GenerateResourceUsePreserializedResources>true</GenerateResourceUsePreserializedResources>
+    <RunAnalyzersDuringBuild>false</RunAnalyzersDuringBuild>
+    <PackageId>karambaConnect</PackageId>
+    <AssemblyName>karambaConnect</AssemblyName>
+    <RootNamespace>karambaConnect</RootNamespace>
+    <DefineConstants>TRACE;karamba1</DefineConstants>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Grasshopper" Version="7.10.21256.17001" IncludeAssets="compile;build" />
+    <PackageReference Include="STBDotNet" Version="0.2.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Common\KarambaConnectCommon.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Reference Include="KarambaCommon">
+      <HintPath>..\Common\Library\RH7_Karamba1\karambaCommon.dll</HintPath>
+      <Private>false</Private>
+    </Reference>
+    <Reference Include="KarambaGH">
+      <HintPath>..\Common\Library\RH7_Karamba1\karamba.gha</HintPath>
+      <Private>false</Private>
+    </Reference>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Include="..\RH7K2\**\*.cs" Exclude="..\RH7K2\karambaConnectRH7K2Info.cs" />
+    <Compile Remove="..\RH7K2\obj\**" />
+  </ItemGroup>
+
+  <PropertyGroup Condition="$(Configuration) == 'Debug' AND $([MSBuild]::IsOSPlatform(Windows))">
+    <StartProgram>C:\Program Files\Rhino 7\System\Rhino.exe</StartProgram>
+    <StartArguments></StartArguments>
+    <StartAction>Program</StartAction>
+  </PropertyGroup>
+
+</Project>
